Handle stopping error in the ApplicationLabel component
