+++ frontend/user/vue-components/ApplicationLabel.vue
@@ -1,5 +1,15 @@
 <template>
   <ul v-if="currentApp !== null" class="nav navbar-nav">
+    <!-- Error dialog -->
+    <confirm-dialog v-if="stoppingError.show"
+    :title="'Error when stopping ' + stoppingError.appName"
+    :okCallback="closeDialog">
+      <div class="alert alert-danger">
+        <strong>Code: {{stoppingError.code}}</strong>
+        <span>{{stoppingError.message}}</span>
+      </div>
+    </confirm-dialog>
+
     <li class="dropdown">
       <a href="#" class="dropdown-toggle cust-padding" data-toggle="dropdown" aria-expanded="false">
         <img class="app-icon"
@@ -29,8 +39,15 @@
 
 <script>
   let Vue = require("vuejs");
+  require('toolkit');
 
   module.exports = Vue.extend({
+    data: function() {
+      return {
+        stoppingError: { show: false, appName: '', code: '', message: '' }
+      };
+    },
+
     computed: {
       currentApp: function() {
         return this.model.appList[this.model.selectedIndex] || null;
@@ -38,10 +55,20 @@
     },
 
     methods: {
-      stopApplication: function(index) {
-        // TODO: Handle error
-        this.model.stopApplication(this.model.selectedIndex);
-      }
+      stopApplication: function() {
+        let stoppingAppName = this.$options.filters.appName(
+          this.currentApp.appData.image);
+        this.model.stopApplication(this.model.selectedIndex).fail((error) => {
+          console.log('Fail')
+          this.stoppingError.code = error.code;
+          this.stoppingError.message = error.message;
+          this.stoppingError.appName = stoppingAppName;
+          this.stoppingError.show = true;
+        });
+      },
+      closeDialog: function() {
+        this.stoppingError.show = false;
+      },
     },
   });
 </script>
